requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
jinja2>=3.1.0
python-multipart>=0.0.6
pywebview>=4.0.0
pyinstaller>=6.0.0
bcrypt>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
